feat: created new Exception for delete

+++ src/main/java/com/devsuperior/DSCommerce/services/ProductService.java
@@ -2,12 +2,16 @@ package com.devsuperior.DSCommerce.services;
 
 import com.devsuperior.DSCommerce.DTO.ProductDTO;
 import com.devsuperior.DSCommerce.entities.Product;
+import com.devsuperior.DSCommerce.exceptions.DatabaseException;
 import com.devsuperior.DSCommerce.exceptions.ResourceNotFoundExeception;
 import com.devsuperior.DSCommerce.repositories.ProductRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -25,7 +29,7 @@ public class ProductService {
     @Transactional(readOnly = true)
     public Page<ProductDTO> findAll(Pageable pageable) {
         Page<Product> productList = repository.findAll(pageable);
-        return productList.map(x -> new ProductDTO(x));
+        return productList.map(ProductDTO::new);
         // return productList.map(ProductDTO::new);
     }
     @Transactional()
@@ -37,15 +41,28 @@ public class ProductService {
     }
     @Transactional()
     public ProductDTO update(Long id, ProductDTO productDTO) {
-        Product product = repository.getReferenceById(id);
-        copyDTO(product, productDTO);
-        product = repository.save(product);
-        return new ProductDTO(product);
+        try {
+            Product product = repository.getReferenceById(id);
+            copyDTO(product, productDTO);
+            product = repository.save(product);
+            return new ProductDTO(product);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundExeception("Recurso não foi encontrado");
+        }
+
     }
 
-    @Transactional()
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundExeception("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDTO(Product product, ProductDTO productDTO) {
